+++ py_slides_term/candidates/data.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, asdict
 from typing import List, Set, Dict, Any, Type
 
-from py_slides_term.morphemes import BaseMorpheme, SpaCyMorpheme
+from py_slides_term.tokenizer import BaseMorpheme, SpaCyMorpheme
 from py_slides_term.share.data import Term
