+++ py_slides_term/candidates/filters/term/concatenation/english.py
@@ -1,7 +1,7 @@
 import re
 
 from ..base import BaseEnglishCandidateTermFilter
-from py_slides_term.morphemes import EnglishMorphemeClassifier
+from py_slides_term.tokenizer import EnglishMorphemeClassifier
 from py_slides_term.share.data import Term
 from py_slides_term.share.consts import ALPHABET_REGEX
