+++ py_slides_term/candidates/filters/term/concatenation/japanese.py
@@ -1,7 +1,7 @@
 import re
 
 from ..base import BaseJapaneseCandidateTermFilter
-from py_slides_term.morphemes import JapaneseMorphemeClassifier
+from py_slides_term.tokenizer import JapaneseMorphemeClassifier
 from py_slides_term.share.data import Term
 from py_slides_term.share.consts import HIRAGANA_REGEX, KATAKANA_REGEX, ALPHABET_REGEX
