+++ py_slides_term/candidates/filters/term/propernoun/japanese.py
@@ -1,5 +1,5 @@
 from ..base import BaseJapaneseCandidateTermFilter
-from py_slides_term.morphemes import BaseMorpheme, JapaneseMorphemeClassifier
+from py_slides_term.tokenizer import BaseMorpheme, JapaneseMorphemeClassifier
 from py_slides_term.share.data import Term
