+++ py_slides_term/methods/rankers/flr.py
@@ -4,7 +4,7 @@ from .base import BaseSingleDomainRanker
 from ..rankingdata import FLRRankingData
 from ..data import MethodTermRanking
 from py_slides_term.candidates import DomainCandidateTermList
-from py_slides_term.morphemes import (
+from py_slides_term.tokenizer import (
     JapaneseMorphemeClassifier,
     EnglishMorphemeClassifier,
     BaseMorpheme,
